Refactor auth pages to server/client split
